requirements.txt
streamlit==1.28.0
qrcode==7.4.2
pillow==10.0.0